\name{sglrWebapp}
\alias{sglrWebapp}
\title{
  A function to run a web application to compute boundaries and plotting
  the decision region
}
\description{
  A function to run a shiny web application to compute boundaries and plotting
  the decision region. Menus are provided for a point-and-click interface

}
\usage{
sglrWebapp()
}
\details{
  The web application runs on your local machine. Internet Explorer is
  known to have trouble at times, so it might be best to use Firefox or
  Chrome as the default browser to use this web application.
}
\author{
  Balasubramanian Narasimhan
}

\examples{
\dontrun{
    sglrWebapp()
}
}
